pyproject.toml
[tool.poetry]
name = "shopify-app"
version = "0.1.0"
description = ""
authors = ["Dimitris Kavelidis <dimitris.kave@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.12"
Django = "~5.1.2"
gunicorn = "~23.0.0"
mysqlclient = "~2.2.5"
python-dotenv = "~1.0.1"
djangorestframework = "^3.15.2"

[tool.poetry.group.dev.dependencies]
pytest = "~8.3.3"
pylint-django = "^2.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
